Deleted some shit, moved some shit around, clean up some shit
