+++ .gitignore
@@ -2,6 +2,7 @@
 **/.empty/
 **/__pycache__/
 **/.pytest_cache/
+**/.ruff_cache/
 
 **/*.out
 **/.DS_Store
