+++ deprecated/visualize_activations.py
