+++ examples/tester.py
@@ -1,20 +1,43 @@
 import os
 import sys
 from traceback import print_exc
+from typing import Any
 
 import hydra
 import torch
+from addict import Dict
 from omegaconf import DictConfig, OmegaConf
 from torch import nn
 from tqdm import tqdm
 
 import wandb
 from bioplnn.utils import (
-    AttrDict,
     initialize_criterion,
     initialize_dataloader,
     initialize_model,
 )
+
+
+class AttrDict(Dict):
+    """A non-default version of the `addict.Dict` class that raises a `KeyError`
+    when a key is not found in the dictionary.
+
+    Args:
+        *args: Any positional arguments.
+        **kwargs: Any keyword arguments.
+    """
+
+    def __missing__(self, key: Any):
+        """Override the default behavior of `addict.Dict` to raise a `KeyError`
+        when a key is not found in the dictionary.
+
+        Args:
+            key: The key that was not found.
+
+        Raises:
+            KeyError: Always raised.
+        """
+        raise KeyError(key)
 
 
 def _test(
