--- helpful_commands.sh
@@ -1,1 +1,0 @@
-python examples/activations.py wandb.mode=disabled model=ei1l data=mazes experiment/activation=relu_tanh experiment/dimensions=channels_32 experiment/rectify=baseline
